# docs/swagger.yaml
definitions:
...
  dto.ReqCreateGroup:
    properties:
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
...
  dto.ReqCreateSubGroup:
    properties:
      goods_group_id:
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - goods_group_id
    - name
    type: object
...
  dto.ReqUpdateGroup:
    properties:
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
...
  dto.ReqUpdateSubGroup:
    properties:
      goods_group_id:
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - goods_group_id
    - name
    type: object
...
  dto.RespGroup:
    properties:
      created_at:
        type: string
      group_code:
        description: omit when creating new group
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.RespGroupIndex:
    properties:
      created_at:
        type: string
      group_code:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.RespSubGroup:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      goods_group_id:
        type: string
      goods_group_name:
        type: string
      id:
        type: string
      name:
        type: string
      subgroup_code:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.RespSubGroupIndex:
    properties:
      created_at:
        type: string
      goods_group_id:
        type: string
      goods_group_name:
        type: string
      id:
        type: string
      name:
        type: string
      subgroup_code:
        type: string
      updated_at:
        type: string
    type: object
...
  response.NonPaginationResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  response.PaginationResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Welcome to the API documentation for the Base Template Web Application.
    This comprehensive guide is designed to help developers seamlessly integrate and
    interact with our platform's functionalities. Whether you're building new features,
    enhancing existing ones, or troubleshooting, this documentation provides all the
    necessary resources and information.
  title: Base Template API Documentation
  version: 0.0-beta
paths:
...
  /v1/group:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of goods groups with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Search keyword for filtering by name and group_code
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved groups
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RespGroupIndex'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get list of groups with pagination
      tags:
      - Golongan
    post:
      consumes:
      - application/json
      description: Create a new goods group with provided name
      parameters:
      - description: Group creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created group
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespGroup'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - Golongan
  /v1/group/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing goods group by ID
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted group
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Delete group
      tags:
      - Golongan
    get:
      consumes:
      - application/json
      description: Retrieve a single goods group by its ID
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved group
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespGroup'
              type: object
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get group by ID
      tags:
      - Golongan
    put:
      consumes:
      - application/json
      description: Update an existing goods group's information
      parameters:
      - description: Group UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated group data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUpdateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated group
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespGroup'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Update group
      tags:
      - Golongan
  /v1/group/export:
    get:
      consumes:
      - application/json
      description: Export goods groups to Excel file (.xlsx) with optional search
        and filter. Same search and filter logic as index but without pagination.
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Search keyword for filtering by name and group_code
        in: query
        name: search
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file with groups data
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Export groups to Excel
      tags:
      - Golongan
...
  /v1/sub-group:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of sub-groups with optional search and
        filters. Supports multiple filter values for subgroup_code, name, and goods_group_id.
        Subgroup codes are in 2-digit string format (e.g., "01", "02", "05", "10").
        Only returns non-deleted sub-groups.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: per_page
        type: integer
      - description: 'Sort column (allowed: id, goods_group_id, subgroup_code, name,
          created_at, updated_at)'
        in: query
        name: sort_by
        type: string
      - description: 'Sort order: asc or desc (default: desc)'
        in: query
        name: sort_order
        type: string
      - description: Search keyword (searches in subgroup_code and name)
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: 'Filter by subgroup codes (multiple values, format: '
        in: query
        items:
          type: string
        name: subgroup_codes
        type: array
      - collectionFormat: csv
        description: Filter by names (multiple values)
        in: query
        items:
          type: string
        name: names
        type: array
      - collectionFormat: csv
        description: Filter by goods group IDs (multiple values, UUIDs)
        in: query
        items:
          type: string
        name: goods_group_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved sub-groups
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RespSubGroupIndex'
                  type: array
              type: object
        "400":
          description: Bad request - invalid query parameters
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get list of sub-groups with pagination
      tags:
      - Sub Golongan
    post:
      consumes:
      - application/json
      description: Create a new sub-group with provided goods_group_id and name. Name
        must be unique within the same goods_group. Subgroup code is automatically
        generated by the database trigger with 2-digit format (e.g., "01", "02", "05",
        "10"). The code is generated sequentially based on the sequence.
      parameters:
      - description: 'Sub-group creation data. Fields: goods_group_id (required, UUID),
          name (required, max 255 chars, uppercase letters only)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateSubGroup'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created sub-group with auto-generated subgroup_code
            in 2-digit format
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespSubGroup'
              type: object
        "400":
          description: Bad request - validation error or duplicate name in goods_group
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Create a new sub-group
      tags:
      - Sub Golongan
  /v1/sub-group/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an existing sub-group by ID. The sub-group will be
        marked as deleted (deleted_at is set) but remains in the database. Requires
        'api.master-data.sub-group.delete' permission.
      parameters:
      - description: Sub-group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully soft deleted sub-group
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Sub-group not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Soft delete sub-group
      tags:
      - Sub Golongan
    get:
      consumes:
      - application/json
      description: Retrieve a single sub-group by its UUID. The response includes
        subgroup_code in 2-digit string format (e.g., "01", "02", "05", "10") and
        goods_group_name to identify which goods group this sub-group belongs to.
        Only returns non-deleted sub-groups.
      parameters:
      - description: Sub-group UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved sub-group with subgroup_code in 2-digit
            format and goods_group_name
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespSubGroup'
              type: object
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Sub-group not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get sub-group by ID
      tags:
      - Sub Golongan
    put:
      consumes:
      - application/json
      description: Update an existing sub-group's information. Name must be unique
        within the same goods_group. Subgroup code cannot be updated as it is auto-generated.
        The response includes the current subgroup_code in 2-digit format.
      parameters:
      - description: Sub-group UUID
        in: path
        name: id
        required: true
        type: string
      - description: 'Updated sub-group data. Fields: goods_group_id (required, UUID),
          name (required, max 255 chars, uppercase letters only). Note: subgroup_code
          is read-only and cannot be updated.'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUpdateSubGroup'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated sub-group with subgroup_code in 2-digit
            format
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespSubGroup'
              type: object
        "400":
          description: Bad request - validation error or duplicate name in goods_group
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Sub-group not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Update sub-group
      tags:
      - Sub Golongan
  /v1/sub-group/export:
    get:
      consumes:
      - application/json
      description: 'Export sub-groups to Excel file (.xlsx) with optional search and
        filter. Same search and filter logic as index but without pagination. Supports
        multiple filter values for each field. Only exports non-deleted sub-groups.
        Excel file includes: Kode Sub Golongan (2-digit format like "01", "02"), Nama
        Sub Golongan, Goods Group ID, Update Date. Requires ''api.master-data.sub-group.export''
        permission.'
      parameters:
      - description: Search keyword (searches in subgroup_code and name)
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: 'Filter by subgroup codes (multiple values, format: '
        in: query
        items:
          type: string
        name: subgroup_codes
        type: array
      - collectionFormat: csv
        description: Filter by names (multiple values)
        in: query
        items:
          type: string
        name: names
        type: array
      - collectionFormat: csv
        description: Filter by goods group IDs (multiple values, UUIDs)
        in: query
        items:
          type: string
        name: goods_group_ids
        type: array
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file (sub-groups.xlsx) with sub-groups data including
            subgroup_code in 2-digit format
          schema:
            type: file
        "400":
          description: Bad request - invalid query parameters
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Export sub-groups to Excel
      tags:
      - Sub Golongan
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token (ex: Bearer eyJhbGciOiJIU....)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"